fixed validation
