+++ controllers/contactsController.js
@@ -1,13 +1,6 @@
-const Joi = require('joi');
 const contactsServise = require("../models/contacts");
 const { HttpError } = require("../helpers");
 const { ctrlWrapper } = require("../decorators");
-
-const addContactSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().required(),
-    phone: Joi.string().required(),
-});
 
 const listContacts = async (_, res) => {
     const result = await contactsServise.listContacts();
@@ -24,10 +17,6 @@ const getContactById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-    const { error } = addContactSchema.validate(req.body);
-    if (error) {
-        throw HttpError(400, `missing required ${error.details[0].context.key} field`);
-    }
     const result = await contactsServise.addContact(req.body);
     res.status(201).json(result);
 };
@@ -44,15 +33,6 @@ const removeContact = async (req, res) => {
 };
 
 const updateContactById = async (req, res) => {
-    if (!Object.keys(req.body).length) {
-        throw HttpError(400, "missing fields");
-    }
-    
-    const { error } = addContactSchema.validate(req.body);
-    if (error) {
-        throw HttpError(400, `missing required ${error.details[0].context.key} field`);
-    }
-
     const { contactId } = req.params;
     const result = await contactsServise.updateContactById(contactId, req.body);
     if (!result) {
