+++ decorators/index.js
@@ -1,5 +1,9 @@
 const ctrlWrapper = require("./ctrlWrapper");
+const validateBody = require("./validateBody");
+const verificateBody = require("./verificateBody");
 
 module.exports = {
     ctrlWrapper,
+    validateBody,
+    verificateBody,
 };
