+++ models/contacts.json
@@ -58,26 +58,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "UT5f-I_rQXDygi6aA_hbm",
-    "name": "allen raymond",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3792"
-  },
-  {
-    "id": "mi-NVf7Nq_U9i8tgiQ-Em",
-    "name": "allen raymond",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3792"
-  },
-  {
-    "id": "AA2yonFBACt93qSaDGNut",
-    "name": "allen raymond",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3792"
-  },
-  {
-    "id": "hSgjKNvblXAEghdHyjThT"
   }
 ]
