+++ routes/api/contacts.js
@@ -1,5 +1,8 @@
 const express = require('express');
 const contactsController = require("../../controllers/contactsController");
+const { validateBody } = require("../../decorators");
+const { verificateBody } = require("../../decorators");
+const schemas = require("../../schemas/contactsSchemas");
 
 const router = express.Router();
 
@@ -7,10 +10,10 @@ router.get('/', contactsController.listContacts);
 
 router.get('/:contactId', contactsController.getContactById);
 
-router.post('/', contactsController.addContact);
+router.post('/', validateBody(schemas.addContactSchema), contactsController.addContact);
 
 router.delete('/:contactId', contactsController.removeContact);
 
-router.put('/:contactId', contactsController.updateContactById);
+router.put('/:contactId', verificateBody(), validateBody(schemas.addContactSchema), contactsController.updateContactById);
 
 module.exports = router;
