+++ schemas/contactsSchemas.js
@@ -1,0 +1,11 @@
+const Joi = require('joi');
+
+const addContactSchema = Joi.object({
+    name: Joi.string().required(),
+    email: Joi.string().required(),
+    phone: Joi.string().required(),
+});
+
+module.exports = {
+    addContactSchema,
+};
